Support compiled regular expressions as a pattern for string fields
This changes the ‘pattern’ argument for string fields in such a way
that compiled regular expressions can be passed as well. This is
useful for specifying flags, such as ‘re.IGNORECASE’.

Internally, a pattern is always stored as compiled regular expression.
The benefits here are that the regular expression is only compiled
once, and that the regular expression itself is syntactically
validated upon field creation, instead of at first use during
validation.

This change also adapts the error messages to print the pattern in a
nice way. Customisation for this is out of scope; see #28.

Updated the docs to reflect the above changes; changed a few typos
while at it.

Closes #42.